move all of ltl::formula to a subdirectory other than instantiation tools
