+++ texada-src/src/formula/texada_atomicprop.cpp
