+++ texada-src/src/formula/texada_constant.cpp
