+++ texada-src/src/formula/texada_constant.h
