+++ texada-src/src/formula/texadatospotformula.cpp
