+++ texada-src/src/formula/texadatospotformula.h
