+++ texada-src/src/formula/texadatospotmapping.cpp
