+++ texada-src/src/formula/texadatospotmapping.h
