+++ texada-src/src/formula/texadatospotrefformula.cpp
